+++ Sources/Decorator/Extensions+Double.swift
@@ -11,7 +11,6 @@ public extension Double {
     var asDate: Date {
         return Date(timeIntervalSince1970: self)
     }
-
     var secondsSince1970: Double {
         return (self / 1000)
     }